Fix LLC consumer to handle segment build fails gracefully (#2499)
* Fix LLC consumer to handle segment build fails gracefully

If segment build fails then we used throw NPE while tyring to delete a null file name.
Fixed the condition, and added a test.
We now handle the condition gracefully, and also set LLC_PARTITION_CONSUMING gauge to 0
so that we can alert on the partition that stopped consumption.

* Modified comment

+++ pinot-core/src/main/java/com/linkedin/pinot/core/data/manager/realtime/LLRealtimeSegmentDataManager.java
@@ -139,7 +139,11 @@ public class LLRealtimeSegmentDataManager extends RealtimeSegmentDataManager {
     }
 
     public void deleteSegmentFile() {
-      FileUtils.deleteQuietly(new File(_segmentFile));
+      // If segment build fails with an exception then we will not be able to create a segment file and
+      // the file name will be null.
+      if (_segmentFile != null) {
+        FileUtils.deleteQuietly(new File(_segmentFile));
+      }
     }
   }
